rename original NeuroMechFly to NeuroMechFlyV0 and moved new NeuroMechFly class to core.py
