+++ flygym/mujoco/__init__.py
@@ -1,1 +1,4 @@
 from .core import Parameters, NeuroMechFly
+from .simulation import Simulation
+from .fly import Fly
+from .camera import Camera
