+++ flygym/mujoco/simulation.py
@@ -1,14 +1,11 @@
-from pathlib import Path
 from typing import Any, Dict, Iterable, List, Optional, Tuple, Union
 
 import gymnasium as gym
 import numpy as np
 from dm_control import mjcf
 from dm_control.utils import transformations
-from flygym.mujoco import preprogrammed, state
 from flygym.mujoco.arena import BaseArena, FlatTerrain
 from flygym.mujoco.camera import Camera
-from flygym.mujoco.core import Parameters
 from flygym.mujoco.fly import Fly
 from gymnasium.core import ObsType, spaces
 
@@ -39,7 +36,7 @@ class Simulation(gym.Env):
     def __init__(
         self,
         flies: Union[Iterable[Fly], Fly],
-        cameras: Union[Iterable[Fly], Camera, None],
+        cameras: Union[Iterable[Camera], Camera, None],
         arena: BaseArena = None,
         timestep: float = 0.0001,
         gravity: Tuple[float, float, float] = (0.0, 0.0, -9.81e3),
@@ -47,9 +44,9 @@ class Simulation(gym.Env):
         """
         Parameters
         ----------
-        flies : Iterable[Fly] or Fly
+        flies : Iterable[flygym.mujoco.fly.Fly] or Fly
             List of flies in the simulation.
-        cameras : Iterable[Fly] or Camera, optional
+        cameras : Iterable[flygym.mujoco.camera.Camera] or Camera, optional
             List of cameras in the simulation. Defaults to the left camera
             of the first fly.
         arena : flygym.mujoco.arena.BaseArena, optional
@@ -496,219 +493,3 @@ class SingleFlySimulation(Simulation):
         key = self.flies[0].name
         obs, reward, terminated, truncated, info = super().step({key: action})
         return obs[key], reward, terminated, truncated, info[key]
-
-
-class NeuroMechFly(SingleFlySimulation):
-    """A NeuroMechFly environment using MuJoCo as the physics engine. This
-    class is a wrapper around the SingleFlySimulation and is provided for
-    backward compatibility.
-
-    Attributes
-    ----------
-    sim_params : flygym.mujoco.Parameters
-        Parameters of the MuJoCo simulation.
-    timestep: float
-        Simulation timestep in seconds.
-    output_dir : Path
-        Directory to save simulation data.
-    arena : flygym.mujoco.arena.BaseArena
-        The arena in which the fly is placed.
-    spawn_pos : Tuple[float, float, float]
-        The (x, y, z) position in the arena defining where the fly will be
-        spawn, in mm.
-    spawn_orientation : Tuple[float, float, float, float]
-        The spawn orientation of the fly in the Euler angle format: (x, y,
-        z), where x, y, z define the rotation around x, y and z in radian.
-    control : str
-        The joint controller type. Can be "position", "velocity", or
-        "torque".
-    init_pose : flygym.state.BaseState
-        Which initial pose to start the simulation from.
-    render_mode : str
-        The rendering mode. Can be "saved" or "headless".
-    actuated_joints : List[str]
-            List of names of actuated joints.
-    contact_sensor_placements : List[str]
-        List of body segments where contact sensors are placed. By
-        default all tarsus segments.
-    detect_flip : bool
-        If True, the simulation will indicate whether the fly has flipped
-        in the ``info`` returned by ``.step(...)``. Flip detection is
-        achieved by checking whether the leg tips are free of any contact
-        for a duration defined in the configuration file. Flip detection is
-        disabled for a period of time at the beginning of the simulation as
-        defined in the configuration file. This avoids spurious detection
-        when the fly is not standing reliably on the ground yet.
-    retina : flygym.mujoco.vision.Retina
-        The retina simulation object used to render the fly's visual
-        inputs.
-    arena_root = dm_control.mjcf.RootElement
-        The root element of the arena.
-    physics: dm_control.mjcf.Physics
-        The MuJoCo Physics object built from the arena's MJCF model with
-        the fly in it.
-    curr_time : float
-        The (simulated) time elapsed since the last reset (in seconds).
-    action_space : gymnasium.core.ObsType
-        Definition of the simulation's action space as a Gym environment.
-    observation_space : gymnasium.core.ObsType
-        Definition of the simulation's observation space as a Gym
-        environment.
-    model : dm_control.mjcf.RootElement
-        The MuJoCo model.
-    vision_update_mask : np.ndarray
-        The refresh frequency of the visual system is often loser than the
-        same as the physics simulation time step. This 1D mask, whose
-        size is the same as the number of simulation time steps, indicates
-        in which time steps the visual inputs have been refreshed. In other
-        words, the visual input frames where this mask is False are
-        repetitions of the previous updated visual input frames.
-    """
-
-    def __init__(
-        self,
-        sim_params: Parameters = None,
-        actuated_joints: List = preprogrammed.all_leg_dofs,
-        contact_sensor_placements: List = preprogrammed.all_tarsi_links,
-        output_dir: Optional[Path] = None,
-        arena: BaseArena = None,
-        xml_variant: Union[str, Path] = "seqik",
-        spawn_pos: Tuple[float, float, float] = (0.0, 0.0, 0.5),
-        spawn_orientation: Tuple[float, float, float] = (0.0, 0.0, np.pi / 2),
-        control: str = "position",
-        init_pose: Union[str, state.KinematicPose] = "stretch",
-        floor_collisions: Union[str, List[str]] = "legs",
-        self_collisions: Union[str, List[str]] = "legs",
-        detect_flip: bool = False,
-    ) -> None:
-        """Initialize a NeuroMechFly environment.
-
-        Parameters
-        ----------
-        sim_params : flygym.mujoco.Parameters
-            Parameters of the MuJoCo simulation.
-        actuated_joints : List[str], optional
-            List of names of actuated joints. By default all active leg
-            DoFs.
-        contact_sensor_placements : List[str], optional
-            List of body segments where contact sensors are placed. By
-            default all tarsus segments.
-        output_dir : Path, optional
-            Directory to save simulation data. If ``None``, no data will
-            be saved. By default None.
-        arena : flygym.mujoco.arena.BaseArena, optional
-            The arena in which the fly is placed. ``FlatTerrain`` will be
-            used if not specified.
-        xml_variant: str or Path, optional
-            The variant of the fly model to use. Multiple variants exist
-            because when replaying experimentally recorded behavior, the
-            ordering of DoF angles in multi-DoF joints depends on how they
-            are configured in the upstream inverse kinematics program. Two
-            variants are provided: "seqik" (default) and "deepfly3d" (for
-            legacy data produced by DeepFly3D, Gunel et al., eLife, 2019).
-            The ordering of DoFs can be seen from the XML files under
-            ``flygym/data/mjcf/``.
-        spawn_pos : Tuple[float, float, float], optional
-            The (x, y, z) position in the arena defining where the fly
-            will be spawn, in mm. By default (0, 0, 0.5).
-        spawn_orientation : Tuple[float, float, float], optional
-            The spawn orientation of the fly in the Euler angle format:
-            (x, y, z), where x, y, z define the rotation around x, y and
-            z in radian. By default (0.0, 0.0, pi/2), which leads to a
-            position facing the positive direction of the x-axis.
-        control : str, optional
-            The joint controller type. Can be "position", "velocity", or
-            "torque", by default "position".
-        init_pose : BaseState, optional
-            Which initial pose to start the simulation from. By default
-            "stretch" kinematic pose with all legs fully stretched.
-        floor_collisions :str
-            Which set of collisions should collide with the floor. Can be
-            "all", "legs", "tarsi" or a list of body names. By default
-            "legs".
-        self_collisions : str
-            Which set of collisions should collide with each other. Can be
-            "all", "legs", "legs-no-coxa", "tarsi", "none", or a list of
-            body names. By default "legs".
-        detect_flip : bool
-            If True, the simulation will indicate whether the fly has
-            flipped in the ``info`` returned by ``.step(...)``. Flip
-            detection is achieved by checking whether the leg tips are free
-            of any contact for a duration defined in the configuration
-            file. Flip detection is disabled for a period of time at the
-            beginning of the simulation as defined in the configuration
-            file. This avoids spurious detection when the fly is not
-            standing reliably on the ground yet. By default False.
-        """
-
-        if sim_params is None:
-            sim_params = Parameters()
-
-        fly = Fly(
-            name="Animat",
-            actuated_joints=actuated_joints,
-            contact_sensor_placements=contact_sensor_placements,
-            xml_variant=xml_variant,
-            spawn_pos=spawn_pos,
-            spawn_orientation=spawn_orientation,
-            control=control,
-            init_pose=init_pose,
-            floor_collisions=floor_collisions,
-            self_collisions=self_collisions,
-            detect_flip=detect_flip,
-            joint_stiffness=sim_params.joint_stiffness,
-            joint_damping=sim_params.joint_damping,
-            actuator_kp=sim_params.actuator_kp,
-            tarsus_stiffness=sim_params.tarsus_stiffness,
-            tarsus_damping=sim_params.tarsus_damping,
-            friction=sim_params.friction,
-            contact_solref=sim_params.contact_solref,
-            contact_solimp=sim_params.contact_solimp,
-            enable_olfaction=sim_params.enable_olfaction,
-            enable_vision=sim_params.enable_vision,
-            render_raw_vision=sim_params.render_raw_vision,
-            vision_refresh_rate=sim_params.vision_refresh_rate,
-            enable_adhesion=sim_params.enable_adhesion,
-            adhesion_force=sim_params.adhesion_force,
-            draw_adhesion=sim_params.draw_adhesion,
-            draw_sensor_markers=sim_params.draw_sensor_markers,
-        )
-
-        if output_dir is not None:
-            Path(output_dir).mkdir(parents=True, exist_ok=True)
-            output_path = Path(output_dir) / "video.mp4"
-        else:
-            output_path = None
-
-        self.output_dir = output_dir
-
-        camera = Camera(
-            fly,
-            window_size=sim_params.render_window_size,
-            play_speed=sim_params.render_playspeed,
-            fps=sim_params.render_fps,
-            camera_id=sim_params.render_camera,
-            timestamp_text=sim_params.render_timestamp_text,
-            play_speed_text=sim_params.render_playspeed_text,
-            draw_contacts=sim_params.draw_contacts,
-            decompose_contacts=sim_params.decompose_contacts,
-            force_arrow_scaling=sim_params.force_arrow_scaling,
-            tip_length=sim_params.tip_length,
-            contact_threshold=sim_params.contact_threshold,
-            draw_gravity=sim_params.draw_gravity,
-            gravity_arrow_scaling=sim_params.gravity_arrow_scaling,
-            align_camera_with_gravity=sim_params.align_camera_with_gravity,
-            camera_follows_fly_orientation=sim_params.camera_follows_fly_orientation,
-            output_path=output_path,
-        )
-
-        super().__init__(
-            fly=fly,
-            cameras=camera,
-            arena=arena,
-            timestep=sim_params.timestep,
-            gravity=sim_params.gravity,
-        )
-
-    def __getattr__(self, item):
-        return getattr(self.fly, item)
